Refactor meeting control bar
