+++ models/love/src/index.ts
@@ -698,10 +698,6 @@ export function createModel (builder: Builder): void {
     searchDisabled: true
   })
 
-  builder.mixin(love.class.MeetingMinutes, core.class.Class, view.mixin.ObjectPanelFooter, {
-    editor: love.component.PanelControlBar
-  })
-
   builder.createDoc(core.class.FullTextSearchContext, core.space.Model, {
     toClass: love.class.MeetingMinutes,
     fullTextSummary: true,
