+++ plugins/love-resources/src/components/meeting/controls/RecordingButton.svelte
@@ -1,0 +1,21 @@
+<script lang="ts">
+
+  import { AccountRole, getCurrentAccount, hasAccountRole } from '@hcengineering/core'
+  import { ModernButton } from '@hcengineering/ui'
+  import { isRecording, isRecordingAvailable, record } from '../../../utils'
+  import love from '../../../plugin'
+  import { lkSessionConnected } from '../../../liveKitClient'
+  import { Room } from '@hcengineering/love'
+
+  export let room: Room
+  export let size: 'large' | 'medium' | 'small' | 'extra-small' | 'min' = 'large'
+</script>
+
+  <ModernButton
+    icon={$isRecording ? love.icon.StopRecord : love.icon.Record}
+    tooltip={{ label: $isRecording ? love.string.StopRecord : love.string.Record }}
+    disabled={!$lkSessionConnected}
+    kind={'secondary'}
+    {size}
+    on:click={() => record(room)}
+  />
