+++ plugins/love-resources/src/components/meeting/widget/TranscriptionTab.svelte
