+++ plugins/love-resources/src/index.ts
@@ -10,8 +10,8 @@ import SelectScreenSourcePopup from './components/SelectScreenSourcePopup.svelte
 import Settings from './components/Settings.svelte'
 import WorkbenchExtension from './components/WorkbenchExtension.svelte'
 import LoveWidget from './components/LoveWidget.svelte'
-import MeetingWidget from './components/widget/MeetingWidget.svelte'
-import WidgetSwitcher from './components/widget/WidgetSwitcher.svelte'
+import MeetingWidget from './components/meeting/widget/MeetingWidget.svelte'
+import WidgetSwitcher from './components/meeting/widget/WidgetSwitcher.svelte'
 import MeetingMinutesPresenter from './components/MeetingMinutesPresenter.svelte'
 import MeetingMinutesSection from './components/MeetingMinutesSection.svelte'
 import EditMeetingMinutes from './components/EditMeetingMinutes.svelte'
@@ -19,7 +19,6 @@ import EditRoom from './components/EditRoom.svelte'
 import FloorAttributePresenter from './components/FloorAttributePresenter.svelte'
 import FloorView from './components/FloorView.svelte'
 import MeetingMinutesTable from './components/MeetingMinutesTable.svelte'
-import PanelControlBar from './components/PanelControlBar.svelte'
 import RoomPresenter from './components/RoomPresenter.svelte'
 import MeetingMinutesDocEditor from './components/MeetingMinutesDocEditor.svelte'
 import MeetingMinutesStatusPresenter from './components/MeetingMinutesStatusPresenter.svelte'
@@ -62,7 +61,6 @@ export default async (): Promise<Resources> => ({
     FloorAttributePresenter,
     FloorView,
     MeetingMinutesTable,
-    PanelControlBar,
     RoomPresenter,
     MeetingMinutesDocEditor,
     MeetingMinutesStatusPresenter,
